refactor: 카테고리별 상품 fetch 함수 분리
